REF: tetst_value_counts
