+++ pandas/tests/series/test_analytics.py
@@ -955,117 +955,6 @@ class TestSeriesAnalytics:
         right.index = pd.MultiIndex.from_tuples(tpls)
         tm.assert_frame_equal(ts.unstack(level=0), right)
 
-    def test_value_counts_datetime(self):
-        # most dtypes are tested in tests/base
-        values = [
-            pd.Timestamp("2011-01-01 09:00"),
-            pd.Timestamp("2011-01-01 10:00"),
-            pd.Timestamp("2011-01-01 11:00"),
-            pd.Timestamp("2011-01-01 09:00"),
-            pd.Timestamp("2011-01-01 09:00"),
-            pd.Timestamp("2011-01-01 11:00"),
-        ]
-
-        exp_idx = pd.DatetimeIndex(
-            ["2011-01-01 09:00", "2011-01-01 11:00", "2011-01-01 10:00"]
-        )
-        exp = pd.Series([3, 2, 1], index=exp_idx, name="xxx")
-
-        s = pd.Series(values, name="xxx")
-        tm.assert_series_equal(s.value_counts(), exp)
-        # check DatetimeIndex outputs the same result
-        idx = pd.DatetimeIndex(values, name="xxx")
-        tm.assert_series_equal(idx.value_counts(), exp)
-
-        # normalize
-        exp = pd.Series(np.array([3.0, 2.0, 1]) / 6.0, index=exp_idx, name="xxx")
-        tm.assert_series_equal(s.value_counts(normalize=True), exp)
-        tm.assert_series_equal(idx.value_counts(normalize=True), exp)
-
-    def test_value_counts_datetime_tz(self):
-        values = [
-            pd.Timestamp("2011-01-01 09:00", tz="US/Eastern"),
-            pd.Timestamp("2011-01-01 10:00", tz="US/Eastern"),
-            pd.Timestamp("2011-01-01 11:00", tz="US/Eastern"),
-            pd.Timestamp("2011-01-01 09:00", tz="US/Eastern"),
-            pd.Timestamp("2011-01-01 09:00", tz="US/Eastern"),
-            pd.Timestamp("2011-01-01 11:00", tz="US/Eastern"),
-        ]
-
-        exp_idx = pd.DatetimeIndex(
-            ["2011-01-01 09:00", "2011-01-01 11:00", "2011-01-01 10:00"],
-            tz="US/Eastern",
-        )
-        exp = pd.Series([3, 2, 1], index=exp_idx, name="xxx")
-
-        s = pd.Series(values, name="xxx")
-        tm.assert_series_equal(s.value_counts(), exp)
-        idx = pd.DatetimeIndex(values, name="xxx")
-        tm.assert_series_equal(idx.value_counts(), exp)
-
-        exp = pd.Series(np.array([3.0, 2.0, 1]) / 6.0, index=exp_idx, name="xxx")
-        tm.assert_series_equal(s.value_counts(normalize=True), exp)
-        tm.assert_series_equal(idx.value_counts(normalize=True), exp)
-
-    def test_value_counts_period(self):
-        values = [
-            pd.Period("2011-01", freq="M"),
-            pd.Period("2011-02", freq="M"),
-            pd.Period("2011-03", freq="M"),
-            pd.Period("2011-01", freq="M"),
-            pd.Period("2011-01", freq="M"),
-            pd.Period("2011-03", freq="M"),
-        ]
-
-        exp_idx = pd.PeriodIndex(["2011-01", "2011-03", "2011-02"], freq="M")
-        exp = pd.Series([3, 2, 1], index=exp_idx, name="xxx")
-
-        s = pd.Series(values, name="xxx")
-        tm.assert_series_equal(s.value_counts(), exp)
-        # check DatetimeIndex outputs the same result
-        idx = pd.PeriodIndex(values, name="xxx")
-        tm.assert_series_equal(idx.value_counts(), exp)
-
-        # normalize
-        exp = pd.Series(np.array([3.0, 2.0, 1]) / 6.0, index=exp_idx, name="xxx")
-        tm.assert_series_equal(s.value_counts(normalize=True), exp)
-        tm.assert_series_equal(idx.value_counts(normalize=True), exp)
-
-    def test_value_counts_categorical_ordered(self):
-        # most dtypes are tested in tests/base
-        values = pd.Categorical([1, 2, 3, 1, 1, 3], ordered=True)
-
-        exp_idx = pd.CategoricalIndex([1, 3, 2], categories=[1, 2, 3], ordered=True)
-        exp = pd.Series([3, 2, 1], index=exp_idx, name="xxx")
-
-        s = pd.Series(values, name="xxx")
-        tm.assert_series_equal(s.value_counts(), exp)
-        # check CategoricalIndex outputs the same result
-        idx = pd.CategoricalIndex(values, name="xxx")
-        tm.assert_series_equal(idx.value_counts(), exp)
-
-        # normalize
-        exp = pd.Series(np.array([3.0, 2.0, 1]) / 6.0, index=exp_idx, name="xxx")
-        tm.assert_series_equal(s.value_counts(normalize=True), exp)
-        tm.assert_series_equal(idx.value_counts(normalize=True), exp)
-
-    def test_value_counts_categorical_not_ordered(self):
-        values = pd.Categorical([1, 2, 3, 1, 1, 3], ordered=False)
-
-        exp_idx = pd.CategoricalIndex([1, 3, 2], categories=[1, 2, 3], ordered=False)
-        exp = pd.Series([3, 2, 1], index=exp_idx, name="xxx")
-
-        s = pd.Series(values, name="xxx")
-        tm.assert_series_equal(s.value_counts(), exp)
-        # check CategoricalIndex outputs the same result
-        idx = pd.CategoricalIndex(values, name="xxx")
-        tm.assert_series_equal(idx.value_counts(), exp)
-
-        # normalize
-        exp = pd.Series(np.array([3.0, 2.0, 1]) / 6.0, index=exp_idx, name="xxx")
-        tm.assert_series_equal(s.value_counts(normalize=True), exp)
-        tm.assert_series_equal(idx.value_counts(normalize=True), exp)
-
     @pytest.mark.parametrize("func", [np.any, np.all])
     @pytest.mark.parametrize("kwargs", [dict(keepdims=True), dict(out=object())])
     @td.skip_if_np_lt("1.15")
